src/orders-api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
azure-monitor-opentelemetry>=1.6.0
